refactor: extract Workflow mixin shared by Fix and Fork
Both verbs duplicated `missing_clone_root`, `build_adapter`,
`resolve_or_fail`, and the `rescue AuthRequired / AdapterError` shell.
Extracted into `GemContribute::CLI::Workflow`, a mixin both classes
include. Each verb still owns `parse_argv`, `execute`, `print_summary`,
and `print_usage_error` — those genuinely differ.

`resolve_target` (formerly two `resolve_or_fail` methods) parameterizes
the verb name for error messages and takes an `allow_owner_repo:` flag
that gates the slash-form bypass — `fix` opts out (it splits its
target on `/` for the issue#), `fork` opts in.

Future `abandon` (per #18 design notes) plugs in by including Workflow
and defining its own three methods.
